use `src` directory
