+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import "../global.css"
+import "../../global.css"
 
 /**
  * @TODO type these!
