+++ src/pages/add/index.tsx
@@ -1,43 +1,36 @@
-import { useRef } from "react"
-import { NextPageContext, NextPage } from "next"
-import Link from "next/link"
+import React, { useRef } from "react"
 import { useRouter } from "next/router"
+import Link from "next/link"
+import { NextPage } from "next"
 import { Formik, Form, Field, FieldArray } from "formik"
+import fetch from "isomorphic-unfetch"
 
-import Layout from "../../../../components/Layout"
-import { sampleFetchWrapper } from "../../../utils/sample-api"
-import { EventDocument } from "../../../../interfaces"
-
-interface Props {
-  event?: EventDocument
-  errors?: string
-}
+import Layout from "../../../components/Layout"
 
 const PLATFORMS = ["Web", "iOS"]
+const INITIAL_VALUES = {
+  name: "",
+  description: "",
+  properties: [] as string[],
+  platforms: [] as string[],
+}
 
-const EditEventById: NextPage<Props> = ({ event }) => {
+const Add: NextPage = () => {
   const router = useRouter()
-  const { id } = router.query
   const ref = useRef<HTMLInputElement>()
   return (
-    <Layout title="Analytics | Edit">
-      <h1>Edit</h1>
+    <Layout title="Analytics | Add">
+      <h1>Add a New Event</h1>
       <Formik
-        initialValues={{
-          name: event?.name,
-          description: event?.description,
-          properties: event?.properties as string[],
-          platforms: event?.platforms as string[],
-        }}
+        initialValues={INITIAL_VALUES}
         onSubmit={async (values, { setSubmitting }) => {
           try {
-            await fetch(`/api/edit/${id}`, {
+            await fetch(`/api/add`, {
               method: "post",
               body: JSON.stringify(values),
             })
-            console.log(values)
             setSubmitting(false)
-            router.push(`/events/${id}`)
+            router.push("/events")
           } catch (err) {
             console.error(err)
             setSubmitting(false)
@@ -133,7 +126,7 @@ const EditEventById: NextPage<Props> = ({ event }) => {
                           type="checkbox"
                           id={`platforms.${index}`}
                           name={`platforms.${index}`}
-                          checked={values.platforms?.includes(_platform)}
+                          checked={values.platforms.includes(_platform)}
                           onChange={(e: any) => {
                             const isChecked: boolean = e.target.checked
                             isChecked
@@ -169,16 +162,4 @@ const EditEventById: NextPage<Props> = ({ event }) => {
   )
 }
 
-EditEventById.getInitialProps = async ({ query }: NextPageContext) => {
-  try {
-    const { id } = query
-    const event = await sampleFetchWrapper(
-      `/api/events/${Array.isArray(id) ? id[0] : id}`
-    )
-    return { event }
-  } catch (err) {
-    return { errors: err.message }
-  }
-}
-
-export default EditEventById
+export default Add
