+++ src/pages/events/[id].tsx
@@ -3,8 +3,8 @@ import Link from "next/link"
 import { useRouter } from "next/router"
 import useSwr from "swr"
 
-import Layout from "../../components/Layout"
-import { EventDocument } from "../../interfaces"
+import Layout from "../../../components/Layout"
+import { EventDocument } from "../../../interfaces"
 
 const EventById: NextPage = () => {
   const router = useRouter()
