+++ src/pages/events/index.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
 import Link from "next/link"
-import Layout from "../../components/Layout"
+import Layout from "../../../components/Layout"
 import useSwr from "swr"
 
-import { EventDocument } from "../../interfaces"
+import { EventDocument } from "../../../interfaces"
 
 function fetcher(url: string) {
   return fetch(url).then(r => r.json())
